+++ transport/src/main/java/io/netty/channel/DefaultChannelId.java
@@ -17,6 +17,7 @@
 package io.netty.channel;
 
 import io.netty.buffer.ByteBufUtil;
+import io.netty.util.internal.MacAddressUtil;
 import io.netty.util.internal.EmptyArrays;
 import io.netty.util.internal.PlatformDependent;
 import io.netty.util.internal.SystemPropertyUtil;
@@ -25,15 +26,7 @@ import io.netty.util.internal.logging.InternalLogger;
 import io.netty.util.internal.logging.InternalLoggerFactory;
 
 import java.lang.reflect.Method;
-import java.net.InetAddress;
-import java.net.NetworkInterface;
-import java.net.SocketException;
-import java.net.UnknownHostException;
 import java.util.Arrays;
-import java.util.Enumeration;
-import java.util.LinkedHashMap;
-import java.util.Map;
-import java.util.Map.Entry;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.regex.Pattern;
 
@@ -47,7 +40,7 @@ final class DefaultChannelId implements ChannelId {
     private static final InternalLogger logger = InternalLoggerFactory.getInstance(DefaultChannelId.class);
 
     private static final Pattern MACHINE_ID_PATTERN = Pattern.compile("^(?:[0-9a-fA-F][:-]?){6,8}$");
-    private static final int MACHINE_ID_LEN = 8;
+    private static final int MACHINE_ID_LEN = MacAddressUtil.MAC_ADDRESS_LENGTH;
     private static final byte[] MACHINE_ID;
     private static final int PROCESS_ID_LEN = 4;
     // Maximal value for 64bit systems is 2^22.  See man 5 proc.
@@ -107,7 +100,7 @@ final class DefaultChannelId implements ChannelId {
         if (machineId == null) {
             machineId = defaultMachineId();
             if (logger.isDebugEnabled()) {
-                logger.debug("-Dio.netty.machineId: {} (auto-detected)", formatAddress(machineId));
+                logger.debug("-Dio.netty.machineId: {} (auto-detected)", MacAddressUtil.formatAddress(machineId));
             }
         }
 
@@ -128,180 +121,15 @@ final class DefaultChannelId implements ChannelId {
     }
 
     private static byte[] defaultMachineId() {
-        // Find the best MAC address available.
-        final byte[] NOT_FOUND = { -1 };
-        byte[] bestMacAddr = NOT_FOUND;
-        InetAddress bestInetAddr = null;
-        try {
-            bestInetAddr = InetAddress.getByAddress(new byte[] { 127, 0, 0, 1 });
-        } catch (UnknownHostException e) {
-            // Never happens.
-            PlatformDependent.throwException(e);
-        }
-
-        // Retrieve the list of available network interfaces.
-        Map<NetworkInterface, InetAddress> ifaces = new LinkedHashMap<NetworkInterface, InetAddress>();
-        try {
-            for (Enumeration<NetworkInterface> i = NetworkInterface.getNetworkInterfaces(); i.hasMoreElements();) {
-                NetworkInterface iface = i.nextElement();
-                // Use the interface with proper INET addresses only.
-                Enumeration<InetAddress> addrs = iface.getInetAddresses();
-                if (addrs.hasMoreElements()) {
-                    InetAddress a = addrs.nextElement();
-                    if (!a.isLoopbackAddress()) {
-                        ifaces.put(iface, a);
-                    }
-                }
-            }
-        } catch (SocketException e) {
-            logger.warn("Failed to retrieve the list of available network interfaces", e);
-        }
-
-        for (Entry<NetworkInterface, InetAddress> entry: ifaces.entrySet()) {
-            NetworkInterface iface = entry.getKey();
-            InetAddress inetAddr = entry.getValue();
-            if (iface.isVirtual()) {
-                continue;
-            }
-
-            byte[] macAddr;
-            try {
-                macAddr = iface.getHardwareAddress();
-            } catch (SocketException e) {
-                logger.debug("Failed to get the hardware address of a network interface: {}", iface, e);
-                continue;
-            }
-
-            boolean replace = false;
-            int res = compareAddresses(bestMacAddr, macAddr);
-            if (res < 0) {
-                // Found a better MAC address.
-                replace = true;
-            } else if (res == 0) {
-                // Two MAC addresses are of pretty much same quality.
-                res = compareAddresses(bestInetAddr, inetAddr);
-                if (res < 0) {
-                    // Found a MAC address with better INET address.
-                    replace = true;
-                } else if (res == 0) {
-                    // Cannot tell the difference.  Choose the longer one.
-                    if (bestMacAddr.length < macAddr.length) {
-                        replace = true;
-                    }
-                }
-            }
-
-            if (replace) {
-                bestMacAddr = macAddr;
-                bestInetAddr = inetAddr;
-            }
-        }
-
-        if (bestMacAddr == NOT_FOUND) {
-            bestMacAddr = new byte[MACHINE_ID_LEN];
+        byte[] bestMacAddr = MacAddressUtil.bestAvailableMac();
+        if (bestMacAddr == null) {
+            bestMacAddr = new byte[MacAddressUtil.MAC_ADDRESS_LENGTH];
             ThreadLocalRandom.current().nextBytes(bestMacAddr);
             logger.warn(
                     "Failed to find a usable hardware address from the network interfaces; using random bytes: {}",
-                    formatAddress(bestMacAddr));
+                    MacAddressUtil.formatAddress(bestMacAddr));
         }
-
-        switch (bestMacAddr.length) {
-            case 6: // EUI-48 - convert to EUI-64
-                byte[] newAddr = new byte[MACHINE_ID_LEN];
-                System.arraycopy(bestMacAddr, 0, newAddr, 0, 3);
-                newAddr[3] = (byte) 0xFF;
-                newAddr[4] = (byte) 0xFE;
-                System.arraycopy(bestMacAddr, 3, newAddr, 5, 3);
-                bestMacAddr = newAddr;
-                break;
-            default: // Unknown
-                bestMacAddr = Arrays.copyOf(bestMacAddr, MACHINE_ID_LEN);
-        }
-
         return bestMacAddr;
-    }
-
-    /**
-     * @return positive - current is better, 0 - cannot tell from MAC addr, negative - candidate is better.
-     */
-    private static int compareAddresses(byte[] current, byte[] candidate) {
-        if (candidate == null) {
-            return 1;
-        }
-
-        // Must be EUI-48 or longer.
-        if (candidate.length < 6) {
-            return 1;
-        }
-
-        // Must not be filled with only 0 and 1.
-        boolean onlyZeroAndOne = true;
-        for (byte b: candidate) {
-            if (b != 0 && b != 1) {
-                onlyZeroAndOne = false;
-                break;
-            }
-        }
-
-        if (onlyZeroAndOne) {
-            return 1;
-        }
-
-        // Must not be a multicast address
-        if ((candidate[0] & 1) != 0) {
-            return 1;
-        }
-
-        // Prefer globally unique address.
-        if ((current[0] & 2) == 0) {
-            if ((candidate[0] & 2) == 0) {
-                // Both current and candidate are globally unique addresses.
-                return 0;
-            } else {
-                // Only current is globally unique.
-                return 1;
-            }
-        } else {
-            if ((candidate[0] & 2) == 0) {
-                // Only candidate is globally unique.
-                return -1;
-            } else {
-                // Both current and candidate are non-unique.
-                return 0;
-            }
-        }
-    }
-
-    /**
-     * @return positive - current is better, 0 - cannot tell, negative - candidate is better
-     */
-    private static int compareAddresses(InetAddress current, InetAddress candidate) {
-        return scoreAddress(current) - scoreAddress(candidate);
-    }
-
-    private static int scoreAddress(InetAddress addr) {
-        if (addr.isAnyLocalAddress() || addr.isLoopbackAddress()) {
-            return 0;
-        }
-        if (addr.isMulticastAddress()) {
-            return 1;
-        }
-        if (addr.isLinkLocalAddress()) {
-            return 2;
-        }
-        if (addr.isSiteLocalAddress()) {
-            return 3;
-        }
-
-        return 4;
-    }
-
-    private static String formatAddress(byte[] addr) {
-        StringBuilder buf = new StringBuilder(24);
-        for (byte b: addr) {
-            buf.append(String.format("%02x:", b & 0xff));
-        }
-        return buf.substring(0, buf.length() - 1);
     }
 
     private static int defaultProcessId() {
